chore: move to assets folder
