+++ front-end/src/assets/data/index.js
@@ -1,0 +1,5 @@
+import collaborators from './collaborators.json';
+import department from './department.json';
+import faqs from './faqs.json'
+
+export { collaborators, department, faqs };
