+++ front-end/src/assets/images/collaborators/value0.png
