+++ front-end/src/assets/images/collaborators/value1.png
