+++ front-end/src/assets/images/collaborators/value2.png
