+++ front-end/src/assets/images/department/index.js
