+++ front-end/src/components/SeekingBanner/SeekingBanner.jsx
@@ -5,13 +5,13 @@ import { background, seekingImage, block1, block2 } from '../../assets/images/co
 export const SeekingBanner = () => {
     return (
         <div className="seeking-banner-section">
-            <img style={{ pointerEvents: "none" }} className="seeking-banner-background" src={background} alt="background" />
+            <img loading="lazy" style={{ pointerEvents: "none" }} className="seeking-banner-background" src={background} alt="background" />
             <div className="seeking-banner-image">
-                <img style={{ pointerEvents: "none" }} src={seekingImage} alt="Collaborators Seeking" />
-                <img style={{ pointerEvents: "none" }} id="seeking-banner-block-2" src={block1} alt="block" />
+                <img loading="lazy" style={{ pointerEvents: "none" }} src={seekingImage} alt="Collaborators Seeking" />
+                <img loading="lazy" style={{ pointerEvents: "none" }} id="seeking-banner-block-2" src={block1} alt="block" />
             </div>
             <div className="seeking-banner-block">
-                <img style={{ pointerEvents: "none" }} id="seeking-banner-block-1" src={block2} alt="block" />
+                <img loading="lazy" style={{ pointerEvents: "none" }} id="seeking-banner-block-1" src={block2} alt="block" />
             </div>
         </div>
     )
