--- front-end/src/mocks/data/achievement.json
@@ -1,3 +1,0 @@
-{
-    
-}
