--- front-end/src/mocks/data/index.js
@@ -1,4 +1,0 @@
-import collaborators from './collaborators.json';
-import department from './department.json'
-
-export { collaborators, department };
