--- front-end/src/mocks/images/collaborators/index.js
@@ -1,1 +1,0 @@
-
