+++ front-end/src/pages/LandingPage/LandingPage.jsx
@@ -8,11 +8,12 @@ import { SeekingProcess } from "../../components/SeekingProcess";
 import { CollaboratorForm } from "../../components/CollaboratorForm";
 import { CountDownTimerV2 } from "../../components/CountDownTimerV2";
 import { Slider } from "../../components/Slider";
+import { Faqs } from "../../components/Faqs";
 
 import { useQuery } from '@tanstack/react-query';
 import { getAchievement } from '../../api/achievement.service';
 
-import { collaborators, department } from '../../mocks/data';
+import { collaborators, department, faqs } from '../../assets/data';
 
 const { form, seekingInfo, coreValues, joinUsInfo, numbersInfo, seekingProcessInfo } = collaborators;
 
@@ -32,6 +33,7 @@ export const LandingPage = () => {
         <Statistics statistics={data}/>
         <CoreValues coreValues={coreValues}/>
         <Slider data={department} activeSlide={2}/>
+        <Faqs faqs={faqs}/>
         <SeekingProcess seekingProcessInfo={seekingProcessInfo}/>
         <hr className="division-bar"/>
         <CountDownTimerV2 targetDate={seekingInfo.date} />
